fix(mail): stay in All Inboxes when switching accounts
Switching accounts from the merged view routed through the account
shortcut, which lands on the new account's inbox — ejecting the reader
from a list that spans every account anyway. The switch logic now lives
in one useAccountSwitch composable (the sidebar submenu and the mobile
profile sheet each had a copy): All Inboxes routes just re-resolve the
active account in place, account-scoped routes swap the accountId param
as before, everything else still goes through the shortcut.
